Rearranging stablize macros
