+++ src/module/macros.ts
@@ -31,8 +31,7 @@ import {
 import { FoundryFlagData, FoundryReg } from "./mm-util/foundry-reg";
 import { is_ref, resolve_dotpath } from "./helpers/commons";
 import { buildSystemHTML } from "./helpers/item";
-import { StabOptions1, StabOptions2 } from "./enums";
-import { applyCollapseListeners, uuid4 } from "./helpers/collapse";
+import { applyCollapseListeners } from "./helpers/collapse";
 import { checkForHit } from "./helpers/automation/targeting";
 import type { AccDiffData, AccDiffDataSerialized, RollModifier } from "./helpers/acc_diff";
 import { is_limited, is_overkill } from "machine-mind/dist/funcs";
@@ -49,6 +48,7 @@ export { encodeMacroData, runEncodedMacro } from "./macros/util"
 export { renderMacroTemplate, renderMacroHTML } from "./macros/render"
 export { prepareActivationMacro } from "./macros/activation"
 export { prepareOverchargeMacro } from "./macros/overcharge"
+export { stabilizeMacro } from "./macros/stabilize"
 export { prepareTechMacro } from "./macros/tech"
 export { prepareTextMacro } from "./macros/text"
 
@@ -1233,52 +1233,6 @@ export function fullRepairMacro(a: string) {
         cancel: {
           icon: '<i class="fas fa-times"></i>',
           label: "No",
-          callback: async () => resolve(false),
-        },
-      },
-      default: "submit",
-      close: () => resolve(false),
-    }).render(true);
-  });
-}
-
-export async function stabilizeMacro(a: string) {
-  // Determine which Actor to speak as
-  let actor = getMacroSpeaker(a);
-  if (!actor) return Promise.reject();
-
-  let template = await renderTemplate(`systems/${game.system.id}/templates/window/promptStabilize.hbs`, {});
-
-  return new Promise<boolean>((resolve, reject) => {
-    new Dialog({
-      title: `STABILIZE - ${actor?.name}`,
-      content: template,
-      buttons: {
-        submit: {
-          icon: '<i class="fas fa-check"></i>',
-          label: "Submit",
-          callback: async dlg => {
-            // Gotta typeguard the actor again
-            if (!actor) return reject();
-
-            let o1 = <StabOptions1>$(dlg).find(".stabilize-options-1:checked").first().val();
-            let o2 = <StabOptions2>$(dlg).find(".stabilize-options-2:checked").first().val();
-
-            let text = await actor.stabilize(o1, o2);
-
-            if (!text) return;
-
-            prepareTextMacro(
-              a,
-              `${actor.name?.capitalize()} HAS STABILIZED`,
-              `${actor.name} has stabilized.<br>${text}`
-            );
-            return resolve(true);
-          },
-        },
-        cancel: {
-          icon: '<i class="fas fa-times"></i>',
-          label: "Cancel",
           callback: async () => resolve(false),
         },
       },
